fix functions structure
