+++ apps/firebase-cloud-functions/src/auth/index.ts
@@ -1,2 +1,4 @@
 // export * from './blocking-functions';
 export * from './crud-user';
+// export * from './fetch-firestore-data';
+
